added some error catchers to get, update, destroy by id that check against id.length before moving forward since mongodb sets a uniform length to 24.

+++ bears/bearController.js
@@ -31,36 +31,61 @@ router
   .route('/:id')
   .get((req, res) => {
     const { id } = req.params;
+    if (id.length !== 24) {
+      res.status(400).send({ error: "The ID entered is too short or too long in length." });
+      return;
+    }
     Bear.findById(id)
       .then(bear => {
+        if (bear === null) {
+          res.status(404).json({ error: "Bear cannot be found with given ID." });
+          return;
+        }
+        // console.log(bear.id.length);
         res.status(200).json(bear);
       })
       .catch(err => {
-        res.status(404).json({ error: "Bear cannot be found with given ID." });
+        res.status(500).json({ error: "Something went terribly wrong!" });
       });
   })
   .delete((req, res) => {
     const { id } = req.params;
+    if (id.length !== 24) {
+      res.status(400).send({ error: "The ID entered is too short or too long in length." });
+      return;
+    }
     Bear.findByIdAndRemove(id)
       .then(bear => {
+        if (bear === null) {
+          res.status(404).json({ error: "Bear cannot be found with given ID." });
+          return;
+        }
         res.json(bear);
       })
       .catch(err => {
-        res.status(404).json({ error: "Bear was not found with given ID." });
+        res.status(500).json({ error: "Something went terribly wrong!" });
       })
   })
   .put((req, res) => {
     const { id } = req.params;
     const { species, latinName } = req.body;
     const updatedBear = { species, latinName };
-
-    Bear.findByIdAndUpdate(id, updatedBear)
-      .then(update => {
-        res.json(updatedBear);
-      })
-      .catch(err => {
-        res.status(404).json({ error: "Bear was not found with given ID." });
-      })
+    if (id.length !== 24) {
+      res.status(400).send({ error: "The ID entered is too short or too long in length." });
+      return;
+    } else {
+      Bear.findByIdAndUpdate(id, updatedBear)
+        .then(update => {
+          if (update === null) {
+            res.status(404).send({ error: "Bear cannot be found with given ID." });
+            return;
+          }
+          res.json(update);
+        })
+        .catch(err => {
+          res.status(500).json({ error: "Something went terribly wrong!" });
+        });
+    };
   });
 
 module.exports = router;
